+++ PocketCartUpgradeSize.cs
@@ -9,7 +9,7 @@ namespace PocketCartPlus
     {
         internal static Value valuePreset = null!;
         internal static readonly float basePriceMultiplier = 4f;
-        public PhysGrabCart Cart;
+        public PhysGrabCart Cart = null!;
         public float chosenScale = 1.25f;
         public Vector3 chosenVector3;
         internal ItemAttributes itemAtts = null!;
@@ -97,43 +97,6 @@ namespace PocketCartPlus
             valuePreset.name = "pocketcartplus_value";
 
             Plugin.Spam($"valuePreset created for cartPlus upgrade with base min price of {HostValues.PlusCartMinPrice.Value} and base max price of {HostValues.PlusCartMaxPrice.Value}");
-        }
-
-        internal static void ShopPatch()
-        {
-            bool shouldAdd = false;
-
-            Item cartPlus = REPOLib.Modules.Items.GetItemByName("Item PCartPlus");
-
-            if (cartPlus == null)
-            {
-                Plugin.Spam($"Item not found!");
-                return;
-            }
-
-            if (HostValues.PlusCartRarity.Value >= Plugin.Rand.Next(0, 100))
-                shouldAdd = true;
-
-            if (!shouldAdd && ShopManager.instance.potentialItems.Contains(cartPlus))
-            {
-                int CountToReplace = ShopManager.instance.potentialItems.Count(i => i.itemAssetName == cartPlus.itemAssetName);
-                Plugin.Spam($"Add-on rarity has determined {cartPlus.itemName} should be removed from the store! Original contains {CountToReplace} of this item");
-                ShopManager.instance.potentialItems.RemoveAll(i => i.itemAssetName == cartPlus.itemAssetName);
-                
-                if(CountToReplace > 0 && ShopManager.instance.potentialItems.Count > 0)
-                {
-                    for(int i = 0; i < CountToReplace; i++)
-                    {
-                        ShopManager.instance.potentialItems.Add(ShopManager.instance.potentialItems[Plugin.Rand.Next(0, ShopManager.instance.potentialItems.Count)]);
-                        Plugin.Spam("Replaced item with another random valid item");
-                    }
-                }  
-            }
-                
-
-            Plugin.Spam($"Rarity determined item is valid to be added to the shop {shouldAdd}");
-            cartPlus.value = valuePreset;
-            Plugin.Spam($"Value preset set for cart small plus!");
         }
     }
 }